Bug 794940: Visible white line in Nexus 7. [r=mfinkle] [needs-clobber]
--HG--
extra : rebase_source : 09b4301502fe5aaf75056d163deb325dabcb79f4

+++ mobile/android/base/MenuButton.java
@@ -30,7 +30,7 @@ public class MenuButton extends ShapedButton {
 
         int width = getMeasuredWidth();
         int height = getMeasuredHeight();
-        float curve = height * 1.125f;
+        int curve = (int) (height * 1.125f);
 
         mPath.reset();
